--- a/Gopkg.toml
+++ b/Gopkg.toml
@@ -46,3 +46,7 @@ required = [
 [[constraint]]
   name = "gopkg.in/src-d/go-git.v4"
   version = "4.6.0"
+
+[[constraint]]
+  name = "github.com/minio/HighwayHash"
+  version = "1.0.0"