… vendor/github.com/minio/HighwayHash/LICENSE
… vendor/github.com/minio/HighwayHash/highwayhash.go
… vendor/github.com/minio/HighwayHash/highwayhashAVX2_amd64.go
… vendor/github.com/minio/HighwayHash/highwayhashAVX2_amd64.s
… vendor/github.com/minio/HighwayHash/highwayhash_amd64.go
… vendor/github.com/minio/HighwayHash/highwayhash_amd64.s
… vendor/github.com/minio/HighwayHash/highwayhash_arm64.go
… vendor/github.com/minio/HighwayHash/highwayhash_arm64.s
… vendor/github.com/minio/HighwayHash/highwayhash_generic.go
… vendor/github.com/minio/HighwayHash/highwayhash_ppc64le.go
… vendor/github.com/minio/HighwayHash/highwayhash_ppc64le.s
… vendor/github.com/minio/HighwayHash/highwayhash_ref.go
… vendor/golang.org/x/sys/cpu/cpu.go
… vendor/golang.org/x/sys/cpu/cpu_arm.go
… vendor/golang.org/x/sys/cpu/cpu_arm64.go
… vendor/golang.org/x/sys/cpu/cpu_gc_x86.go
… vendor/golang.org/x/sys/cpu/cpu_gccgo.c
… vendor/golang.org/x/sys/cpu/cpu_gccgo.go
… vendor/golang.org/x/sys/cpu/cpu_mips64x.go
… vendor/golang.org/x/sys/cpu/cpu_mipsx.go
… vendor/golang.org/x/sys/cpu/cpu_ppc64x.go
… vendor/golang.org/x/sys/cpu/cpu_s390x.go
… vendor/golang.org/x/sys/cpu/cpu_x86.go
… vendor/golang.org/x/sys/cpu/cpu_x86.s