* replace md5 to highwayhash
* optimize HasFilepathPrefix

Result in my project:
runtime.main: 87.51s -> 55.40s
total: 129.54s -> 88.29s

## before:
![profile001](https://user-images.githubusercontent.com/1029249/58904176-f84e0f00-8741-11e9-84b0-00c1c50ce71b.png)

## after:
![profile002](https://user-images.githubusercontent.com/1029249/58904235-1582dd80-8742-11e9-80c9-fb5c06deca93.png)