@orisano 2019-06-04
Bench Result of HasFilepathPrefix.
```
name                                           old time/op    new time/op    delta
HasFilepathPrefix/PathDepth=2,PrefixDepth=1-4     297ns ± 9%     289ns ±21%     ~     (p=0.726 n=10+10)
HasFilepathPrefix/PathDepth=3,PrefixDepth=1-4     298ns ± 3%     270ns ± 6%   -9.16%  (p=0.000 n=10+9)
HasFilepathPrefix/PathDepth=4,PrefixDepth=1-4     332ns ± 4%     289ns ±13%  -13.04%  (p=0.000 n=10+10)
HasFilepathPrefix/PathDepth=5,PrefixDepth=1-4     384ns ±14%     319ns ± 7%  -16.92%  (p=0.000 n=10+10)
HasFilepathPrefix/PathDepth=2,PrefixDepth=2-4     303ns ±14%     337ns ±38%     ~     (p=0.133 n=10+9)
HasFilepathPrefix/PathDepth=3,PrefixDepth=2-4     362ns ±17%     346ns ±13%     ~     (p=0.247 n=10+10)
HasFilepathPrefix/PathDepth=4,PrefixDepth=2-4     427ns ±22%     390ns ±16%     ~     (p=0.086 n=10+10)
HasFilepathPrefix/PathDepth=5,PrefixDepth=2-4     403ns ± 9%     347ns ± 6%  -13.88%  (p=0.000 n=9+9)
HasFilepathPrefix/PathDepth=2,PrefixDepth=3-4     342ns ± 7%     353ns ±13%     ~     (p=0.530 n=9+9)
HasFilepathPrefix/PathDepth=3,PrefixDepth=3-4     413ns ±15%     353ns ± 8%  -14.39%  (p=0.001 n=10+10)
HasFilepathPrefix/PathDepth=4,PrefixDepth=3-4     531ns ±28%     383ns ± 3%  -27.77%  (p=0.000 n=9+10)
HasFilepathPrefix/PathDepth=5,PrefixDepth=3-4     485ns ±29%     424ns ±19%  -12.53%  (p=0.009 n=10+10)

name                                           old alloc/op   new alloc/op   delta
HasFilepathPrefix/PathDepth=2,PrefixDepth=1-4     80.0B ± 0%     80.0B ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=3,PrefixDepth=1-4     96.0B ± 0%     80.0B ± 0%  -16.67%  (p=0.000 n=10+10)
HasFilepathPrefix/PathDepth=4,PrefixDepth=1-4      112B ± 0%       80B ± 0%  -28.57%  (p=0.000 n=10+10)
HasFilepathPrefix/PathDepth=5,PrefixDepth=1-4      128B ± 0%       80B ± 0%  -37.50%  (p=0.000 n=10+10)
HasFilepathPrefix/PathDepth=2,PrefixDepth=2-4     96.0B ± 0%    112.0B ± 0%  +16.67%  (p=0.000 n=10+10)
HasFilepathPrefix/PathDepth=3,PrefixDepth=2-4      112B ± 0%      112B ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=4,PrefixDepth=2-4      128B ± 0%      112B ± 0%  -12.50%  (p=0.000 n=10+10)
HasFilepathPrefix/PathDepth=5,PrefixDepth=2-4      144B ± 0%      112B ± 0%  -22.22%  (p=0.000 n=10+10)
HasFilepathPrefix/PathDepth=2,PrefixDepth=3-4      112B ± 0%      144B ± 0%  +28.57%  (p=0.000 n=10+10)
HasFilepathPrefix/PathDepth=3,PrefixDepth=3-4      128B ± 0%      144B ± 0%  +12.50%  (p=0.000 n=10+10)
HasFilepathPrefix/PathDepth=4,PrefixDepth=3-4      144B ± 0%      144B ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=5,PrefixDepth=3-4      160B ± 0%      144B ± 0%  -10.00%  (p=0.000 n=10+10)

name                                           old allocs/op  new allocs/op  delta
HasFilepathPrefix/PathDepth=2,PrefixDepth=1-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=3,PrefixDepth=1-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=4,PrefixDepth=1-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=5,PrefixDepth=1-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=2,PrefixDepth=2-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=3,PrefixDepth=2-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=4,PrefixDepth=2-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=5,PrefixDepth=2-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=2,PrefixDepth=3-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=3,PrefixDepth=3-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=4,PrefixDepth=3-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
HasFilepathPrefix/PathDepth=5,PrefixDepth=3-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
```