@tejal29 2019-09-17
Looks like this failed
```
Building images for Dockerfile Dockerfile_test_meta_arg
Error building images: Failed to build image gcr.io/kaniko-test/docker-dockerfile_test_registry with kaniko command "[docker run -v /home/kbuilder/.config/gcloud:/root/.config/gcloud -v /tmpfs/tmp/622634064:/kaniko/benchmarks -v /usr/local/go/src/github.com/GoogleContainerTools/kaniko/integration:/workspace -e BENCHMARK_FILE=false executor-image -f /workspace/dockerfiles/Dockerfile_test_registry -d gcr.io/kaniko-test/kaniko-dockerfile_test_registry  -c /workspace]": exit status 1 
```